requirements.txt
Flask
Flask-WTF
Flask-Limiter
python-dotenv
